tunas/requirements.txt
# YOLOv5 v6 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
pandas>=1.1.4
seaborn>=0.11.0
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.64.0
addict>=2.4.0
pycocotools>=2.0
thop  # FLOPs computation
pytorch_quantization>=2.1.1
tensorboard>=2.7.0


